Pequeño cambio para los eventos en el backend

+++ backend/crud/event.py
@@ -1,37 +1,62 @@
 from sqlalchemy.orm import Session
 from models.event import Event
 from schemas.event import EventCreate, EventUpdate
-from datetime import datetime
-import pytz
-
-
+from datetime import datetime, timedelta
 
 # Dar de alta al usuario
 def create_event(db: Session,  event: EventCreate):
-    db_event = Event(event_name=event.event_name, event_type=event.event_type, event_edition=event.event_edition,category=event.category,description=event.description, location=event.location, celebration_date=event.celebration_date, capacity=event.capacity)
+    db_event = Event(user_id=event.user_id,event_name=event.event_name, event_type=event.event_type, event_edition=event.event_edition,category=event.category,event_description=event.event_description, location=event.location, celebration_date=event.celebration_date, capacity=event.capacity)
     db.add(db_event)
     db.commit() 
-    return "Evento creado"
+    return db_event
 
 def find_event_by_name(db: Session, event_name: str):
     return db.query(Event).filter(Event.event_name == event_name).first()
+
+def find_event_by_userId(db: Session, user_id: int):
+    return db.query(Event).filter(Event.user_id == user_id).all()
+
+def find_all_event(db: Session):
+    return db.query(Event).all()
+
 
 def remove_event(db: Session,event_name: str):
     db_event = find_event_by_name(db=db, event_name=event_name)
     if not db_event:
        return False   
+    
+    actual_date = datetime.now()
+    diferencia = db_event.celebration_date - actual_date
+
+    if actual_date > db_event.celebration_date:
+        return False
+    
+    if actual_date == db_event.celebration_date:
+        return False
+    
+    if diferencia < timedelta(days=3):
+        return False
+
     db.delete(db_event)
     db.commit()
-    return "Evento eliminado"
+    return "Event removed"
        
 def change_event(db: Session, event: EventUpdate):
     db_event =  find_event_by_name(db=db, event_name=event.event_name)
     if not db_event:
        return False
     
-    celebration_datetime = datetime.now()
-    local_tz = pytz.timezone('Europe/Madrid')
-    celebration_datetime = local_tz.localize(celebration_datetime)
+    actual_date = datetime.now()
+    diferencia = db_event.celebration_date - actual_date
+
+    if actual_date > db_event.celebration_date:
+        return False
+    
+    if actual_date == db_event.celebration_date:
+        return False
+    
+    if diferencia < timedelta(days=3):
+        return False
 
     update_data = event.model_dump(exclude_unset=True)
